requirements.txt
flask
flask_session
sqlite3
werkzeug.security
requests
